requirements.txt
python-dotenv==1.0.0
pytz==2023.3.post1
pythorhead==0.15.6
requests==2.31.0
python-dateutil==2.8.2